# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Sercovir Geopolitical Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /statements:
    get:
      operationId: listStatements
      tags: [statements]
      summary: List all public statements, tweets, posts
      parameters:
        - name: leaderId
          in: query
          schema:
            type: integer
        - name: platform
          in: query
          schema:
            type: string
        - name: topic
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Statements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderStatement"
    post:
      operationId: createStatement
      tags: [statements]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStatementBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderStatement"
  /statements/{id}:
    get:
      operationId: getStatement
      tags: [statements]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderStatement"
    delete:
      operationId: deleteStatement
      tags: [statements]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── PARLIAMENTARY DISCUSSIONS ─────────────────────────────────────────────
  /parliamentary:
    get:
      operationId: listParliamentary
      tags: [parliamentary]
      summary: List parliamentary discussions and debates
      parameters:
        - name: country
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Discussions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParliamentaryDiscussion"
    post:
      operationId: createParliamentary
      tags: [parliamentary]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateParliamentaryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParliamentaryDiscussion"
  /parliamentary/{id}:
    get:
      operationId: getParliamentary
      tags: [parliamentary]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Discussion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParliamentaryDiscussion"
    put:
      operationId: updateParliamentary
      tags: [parliamentary]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateParliamentaryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParliamentaryDiscussion"
    delete:
      operationId: deleteParliamentary
      tags: [parliamentary]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── LEGISLATION ───────────────────────────────────────────────────────────
  /legislation:
    get:
      operationId: listLegislation
      tags: [legislation]
      summary: List laws in discussion and new legislation
      parameters:
        - name: country
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Legislation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Legislation"
    post:
      operationId: createLegislation
      tags: [legislation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLegislationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Legislation"
  /legislation/{id}:
    get:
      operationId: getLegislation
      tags: [legislation]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Legislation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Legislation"
    put:
      operationId: updateLegislation
      tags: [legislation]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLegislationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Legislation"
    delete:
      operationId: deleteLegislation
      tags: [legislation]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── CRIMINAL CASES ────────────────────────────────────────────────────────
...
  /treaties:
    get:
      operationId: listTreaties
      tags: [treaties]
      summary: List international treaties
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Treaties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Treaty"
    post:
      operationId: createTreaty
      tags: [treaties]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTreatyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treaty"
  /treaties/{id}:
    get:
      operationId: getTreaty
      tags: [treaties]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Treaty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treaty"
    put:
      operationId: updateTreaty
      tags: [treaties]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTreatyBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treaty"
    delete:
      operationId: deleteTreaty
      tags: [treaties]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── MEDIA EVENTS ──────────────────────────────────────────────────────────
  /media-events:
    get:
      operationId: listMediaEvents
      tags: [media-events]
      summary: List major global incidents and news events
      parameters:
        - name: country
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Media events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaEvent"
    post:
      operationId: createMediaEvent
      tags: [media-events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMediaEventBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaEvent"
  /media-events/{id}:
    get:
      operationId: getMediaEvent
      tags: [media-events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Media event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaEvent"
    delete:
      operationId: deleteMediaEvent
      tags: [media-events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── COUNTRY INTELLIGENCE ──────────────────────────────────────────────────
...
components:
  schemas:
...
    LeaderStatement:
      type: object
      properties:
        id:
          type: integer
        leaderId:
          type: integer
        leaderName:
          type: string
        country:
          type: string
        platform:
          type: string
        content:
          type: string
        sentiment:
          type: string
        topic:
          type: string
        relatedCountries:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
        engagement:
          type: integer
        isControversial:
          type: boolean
      required: [id, leaderId, content, platform, timestamp]
    CreateStatementBody:
      type: object
      properties:
        leaderId:
          type: integer
        platform:
          type: string
        content:
          type: string
        sentiment:
          type: string
        topic:
          type: string
        relatedCountries:
          type: array
          items:
            type: string
        engagement:
          type: integer
        isControversial:
          type: boolean
      required: [leaderId, content, platform]
    ParliamentaryDiscussion:
      type: object
      properties:
        id:
          type: integer
        country:
          type: string
        countryCode:
          type: string
        chamber:
          type: string
        topic:
          type: string
        description:
          type: string
        status:
          type: string
        date:
          type: string
        outcome:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        votesFor:
          type: integer
        votesAgainst:
          type: integer
        abstentions:
          type: integer
        significance:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, country, chamber, topic, status]
    CreateParliamentaryBody:
      type: object
      properties:
        country:
          type: string
        countryCode:
          type: string
        chamber:
          type: string
        topic:
          type: string
        description:
          type: string
        status:
          type: string
        date:
          type: string
        outcome:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        votesFor:
          type: integer
        votesAgainst:
          type: integer
        abstentions:
          type: integer
        significance:
          type: string
      required: [country, chamber, topic, status]
    Legislation:
      type: object
      properties:
        id:
          type: integer
        country:
          type: string
        countryCode:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        category:
          type: string
        proposedDate:
          type: string
        enactedDate:
          type: string
        proposedBy:
          type: string
        supportingParties:
          type: array
          items:
            type: string
        opposingParties:
          type: array
          items:
            type: string
        publicOpinion:
          type: string
        impact:
          type: string
        controversyLevel:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, country, title, status, category]
    CreateLegislationBody:
      type: object
      properties:
        country:
          type: string
        countryCode:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        category:
          type: string
        proposedDate:
          type: string
        enactedDate:
          type: string
        proposedBy:
          type: string
        supportingParties:
          type: array
          items:
            type: string
        opposingParties:
          type: array
          items:
            type: string
        publicOpinion:
          type: string
        impact:
          type: string
        controversyLevel:
          type: string
      required: [country, title, status, category]
...
    Treaty:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        signatories:
          type: array
          items:
            type: string
        status:
          type: string
        signedDate:
          type: string
        effectiveDate:
          type: string
        description:
          type: string
        significance:
          type: string
        relatedConflicts:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, type, status]
    CreateTreatyBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        signatories:
          type: array
          items:
            type: string
        status:
          type: string
        signedDate:
          type: string
        effectiveDate:
          type: string
        description:
          type: string
        significance:
          type: string
        relatedConflicts:
          type: array
          items:
            type: integer
      required: [title, type, status]
    MediaEvent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        country:
          type: string
        countryCode:
          type: string
        category:
          type: string
        severity:
          type: string
        description:
          type: string
        date:
          type: string
        source:
          type: string
        isVerified:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, country, category, severity]
    CreateMediaEventBody:
      type: object
      properties:
        title:
          type: string
        country:
          type: string
        countryCode:
          type: string
        category:
          type: string
        severity:
          type: string
        description:
          type: string
        date:
          type: string
        source:
          type: string
        isVerified:
          type: boolean
        tags:
          type: array
          items:
            type: string
      required: [title, country, category, severity]
...